ApiChangeLogs , ApiChangeLog 어노테이션 개선 필요 : version : 버전 2.0.0 으로 수정 및 suhsaechan.kr 로 도메인명 수정 https://github.com/Cassiiopeia/suh-api-log/issues/14
