+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
     id 'maven-publish'
 }
 
-group = 'me.suhsaechan'
+group = 'kr.suhsaechan'
 version = project.findProperty('version') ?: '1.0.4' 
 sourceCompatibility = '17'
