+++ src/main/java/kr/suhsaechan/suhapilog/SuhApiLogApplication.java
@@ -1,0 +1,13 @@
+package kr.suhsaechan.suhapilog;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SuhApiLogApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(SuhApiLogApplication.class, args);
+  }
+
+}
