+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-me.suhsaechan.suhapilog.config.SuhApiLogAutoConfiguration
+kr.suhsaechan.suhapilog.config.SuhApiLogAutoConfiguration
