+++ src/test/java/kr/suhsaechan/suhapilog/SuhApiLogApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.suhsaechan.suhapilog;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SuhApiLogApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
