+++ src/test/java/kr/suhsaechan/suhapilog/config/SwaggerConfig.java
@@ -1,0 +1,37 @@
+package kr.suhsaechan.suhapilog.config;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import io.swagger.v3.oas.models.security.SecurityScheme;
+import java.util.List;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class SwaggerConfig implements WebMvcConfigurer {
+
+  @Bean
+  public OpenAPI openAPI() {
+    SecurityScheme apiKey = new SecurityScheme()
+        .type(SecurityScheme.Type.HTTP)
+        .in(SecurityScheme.In.HEADER)
+        .name("Authorization")
+        .scheme("bearer")
+        .bearerFormat("JWT");
+
+    SecurityRequirement securityRequirement = new SecurityRequirement()
+        .addList("Bearer Token");
+
+    return new OpenAPI()
+        .components(new Components().addSecuritySchemes("Bearer Token", apiKey))
+        .addSecurityItem(securityRequirement)
+        .servers(List.of(
+                new io.swagger.v3.oas.models.servers.Server()
+                    .url("http://localhost:8080")
+                    .description("로컬 서버")
+            )
+        );
+  }
+}
