--- src/test/java/me/suhsaechan/suhapilog/SuhApiLogApplicationTests.java
@@ -1,13 +1,0 @@
-package me.suhsaechan.suhapilog;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SuhApiLogApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
